Adjust clock face digits to display accurate values
The clock now shows the correct time relative to the suns position
but a side effect of this is that 0 is no longer at the top and so on.

I would like 0 to always be on the top and let the digits spread according
to the suns position.
This would mean the distance between digits would variate, but 0 would
 always be on top